cocomaps/conda_requirements.txt
python==3.11
numpy==1.26.4
pandas==2.2.2
scipy==1.14.0
bidict==0.23.1
cctbx-base==2024.5
biopython==1.83
requests==2.32.3
